% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-bib.R
\name{gen_bib_url}
\alias{gen_bib_url}
\title{Generate bib citations for URLs for some given strings}
\usage{
gen_bib_url(handle, url, title = handle, date = Sys.Date(), clipboard = FALSE)
}
\arguments{
\item{handle}{character - the name of the handle}

\item{url}{character - the URL}

\item{title}{character - the title for the bib reference. Default is to use
the handle}

\item{date}{character - the date accessed. Default is current date}

\item{clipboard}{logical - do you want to use the clipboard? Default to FALSE}
}
\value{
string of the bib handle
}
\description{
Generate bib citations for URLs for some given strings
}
\examples{
gen_bib_url(handle = "wikipedia",
            url = "https://en.wikipedia.org/wiki/Main_Page")
}
